pyproject.toml
[project]
name = "asteroid-smash"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Brandon Corfman", email = "bcorfman@fastmail.fm" }
]
dependencies = [
    "numpy>=1.26.4",
    "cocos2d>=0.6.10",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.22.2",
    "pytest>=8.1.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/asteroid_smash"]
